New target/simulation source code file structure
