+++ scripts/Makefile
@@ -1,2 +1,2 @@
 all:
-	gcc simulation_client.c -o simulation_client ../src/search.c -I../src/ -lzmq
+	gcc -DMMSIM_SIMULATION simulation_client.c -o simulation_client ../src/search.c ../src/solve.c -I../src/ ../src/simulation/move.c -I../src/simulation/ -lzmq
