+++ src/simulation/detection.h
@@ -1,0 +1,8 @@
+#ifndef __DETECTION_H
+#define __DETECTION_H
+
+#include "search.h"
+
+extern struct walls_around read_walls(void);
+
+#endif /* __DETECTION_H */
