+++ src/simulation/move.c
@@ -1,0 +1,10 @@
+#include "move.h"
+
+bool collision_detected(void)
+{
+	return false;
+}
+
+void move(enum step_direction direction)
+{
+}
